Promote core-region point-cloud distance to production; fix TF-stamp bug blocking PointCloud2
Core-region median (was diagnostic-only) now used for distance_m/x_cam/
y_cam/z_cam, falling back to full-box when the eroded region doesn't clear
POINTCLOUD_MIN_VALID_POINTS. Ruler benchmark confirms real improvement:
0.30m -> 0.28m (was 0.43m), 0.50m -> 0.47m (was 0.50m), 1.00m -> 0.89-0.91m
stable (was 0.92-1.29m swinging wildly) -- edge-bleeding bias from the
looser detection box is gone.

Root-caused why /duck_pointcloud_region never rendered in RViz2 despite
correct topic/TF wiring: on_odom copied msg.header.stamp into the odom->
base_link TF broadcast, but the ESP32/micro-ROS Odometry message's stamp
is always zero -- pinning that TF at "time 0" forever and breaking any
tf2 lookup at the current time. Sphere Markers (placed directly in the
odom frame, no TF lookup needed) masked this for days. Fixed to stamp
with the Jetson's own wall clock.

Code review of this diff caught two more real issues: the debug point
cloud publish was still using the full-box mask unconditionally (would
have shown stale, misleading data for exactly the elongation check this
fix was meant to support), and the full-box/core-region median math was
duplicated -- both fixed (shared _median_xyz_distance helper, debug cloud
now tracks whichever region actually produced the reported distance).

Also fixes a float32/float64 dtype crash in stereo_calibrate_fisheye.py
(cv2.cornerSubPix returns float32, cv2.fisheye.stereoCalibrate requires
float64 throughout) -- used it to rule out lens-model mismatch as the
cause of the residual ~5-11% distance bias: fisheye and pinhole models
independently converge to baseline within 0.2% of each other, so the
calibration itself is not the source of that remaining bias.

+++ jetson/calibration/stereo_calibrate_fisheye.py
@@ -44,7 +44,12 @@ def find_corners(img_paths):
 
         corners = cv2.cornerSubPix(gray, corners, (11, 11), (-1, -1), SUBPIX_CRITERIA)
         objpoints.append(objp)
-        imgpoints.append(corners.reshape(1, -1, 2))
+        # cv2.cornerSubPix returns float32 -- cv2.fisheye.stereoCalibrate's
+        # binding is strict about matching float64 throughout (objp above
+        # is float64), and mixing the two throws a bare "different types"
+        # arithm_op error deep inside the C++ call, not a helpful message
+        # pointing at this line (found live 2026-08-29).
+        imgpoints.append(corners.reshape(1, -1, 2).astype(np.float64))
         used_paths.append(path)
 
     return objpoints, imgpoints, used_paths, image_size
